whatsapp-invoice-bot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
twilio==9.3.3
anthropic==0.34.2
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
